Cleaned up code, added log-in/log-out, and made /all work correctly
I went through the pre-existing code given to the class and cleaned it up to better my understanding of how it works. After I got a basic idea of how it works, I was able to add primitive log-in and log-out functionality as well as adding an /all page which has active hyperlinks for each item listed.