Fixed issues with the edit functionality
I fixed the issues related to the edit function. Initially, the issues I was having were due to the desired page-to-be edited name was not being passed into the edit webpage. This made is so the page would not save the edits; upon saving the server could not load an empty route and would return an error page.